requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-google-genai>=1.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
rank_bm25>=0.2.2